Rearranged package structure; added __init__ s
The package is rearranged for future proofing. Currently the package has
NumInt module (which handles numerical integration related functions) and
AuxFuncs module (which handles auxiliary and ultility functions that might
be usefult across modules, params_container etc.). The __init__ for all
folders are written and __all__ added so that the package has proper behaviour.

To-do:
    -other integration method not implemented
    -add a tarred config files that will be sync to github